backend/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
idna==3.7
oauthlib==3.2.2
psycopg==3.1.18
psycopg-binary==3.1.18
python-dotenv==1.0.1
requests==2.31.0
requests-oauthlib==2.0.0
sqlparse==0.4.4
typing_extensions==4.11.0
urllib3==2.2.1
